report/requirements.txt
fastapi
uvicorn
requests
pandas
openpyxl
reportlab
sqlalchemy
mysql-connector-python
python-dotenv
XlsxWriter
